# ~100x perf improvement of `micromatch.not` on large inputs by using `Set.prototype.has` over `Array.prototype.includes`