@jonschlinkert 2022-03-18
Wow, this is fantastic! This is how you do a PR. I just glanced at the code and I'm 100% on board with this change. Thank you so much for the time you spent on this, for isolating the problem, creating a solution, and taking the time to provide a detailed explanation. 

I'm about to hop in the car, and won't be back online to merge this until maybe later this weekend. But I definitely approve this and would be happy with someone else merging and publishing a release, if you want to do it before I get to it. 

@joyceerhl thank you!